3_agentic_automation_with_mcp/requirements.txt
langchain-mcp-adapters==0.1.9